+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added Models for each data type
 * Updated getRecords to return Records instead of Maps
 * Added writeData method
+* Added deleteRecords methods
 ## [1.2.3]
 * Upgrade to `alpha11` [#8](https://github.com/duynguyen242/flutter_health_connect/pull/8) by [aljkor](https://github.com/aljkor)
 ## [1.2.2]